backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.1.0
APScheduler==3.10.4
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1